% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode_context}
\alias{encode_context}
\title{Generate LLM-friendly context for the encodeproject.org package}
\usage{
encode_context()
}
\value{
Character string (invisibly), also printed
}
\description{
Generate LLM-friendly context for the encodeproject.org package
}
