% Generated by build script
\name{encode_files}
\alias{encode_files}
\title{List files for an ENCODE experiment}
\usage{
encode_files()
}
\arguments{
\item{accession}{ENCODE experiment accession}
\item{file_format}{Filter by format: "bam", "bigWig", "fastq", "bed", etc.}
}
\value{tibble: accession, file_format, output_type, assembly, file_size, href, status}
\description{List files for an ENCODE experiment}
