% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode_files}
\alias{encode_files}
\title{List files for an ENCODE experiment}
\usage{
encode_files(accession, file_format = NULL)
}
\arguments{
\item{accession}{ENCODE experiment accession}

\item{file_format}{Filter by format: "bam", "bigWig", "fastq", "bed", etc.}
}
\value{
tibble: accession, file_format, output_type, assembly, file_size, href, status
}
\description{
List files for an ENCODE experiment
}
